requirements.txt
Flask==2.3.3
gunicorn==21.2.0
dateparser==1.2.2
duckdb==1.3.2
matplotlib
mistralai==1.9.1
nltk
numpy==1.26.4
openpyxl==3.1.5
pandas==2.3.1
pdf2image==1.17.0
Pillow
PyPDF2==3.0.1
pypdf==6.0.0
pytesseract==0.3.13
python-dateutil
python-docx==1.2.0
reportlab==4.4.3
scikit-learn==1.7.0
sentence-transformers==5.0.0
textract
tqdm
